+++ src/app.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
+import { HttpResponseModule } from '@wlisfes/chat-web-base-schema/interceptor'
 import { validateEnvironment } from '@/config/environment'
 import { ServiceConfigModule } from '@/modules/config/config.module'
 import { GatewayModule } from '@/modules/gateway/gateway.module'
@@ -7,6 +8,7 @@ import { NacosModule } from '@/modules/nacos/nacos.module'
 
 @Module({
     imports: [
+        HttpResponseModule,
         ConfigModule.forRoot({
             isGlobal: true,
             cache: true,